backend/requirements.txt
python-dotenv
crewai
langchain-groq
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask
flask-cors
langchain
requests
Flask-Session